MOVE; auth.service, auth.controller test files
to test folder
